modify - awardRouter 경로 수정, 사용자 validation 기능 추가

+++ back/src/routers/awardRouter.js
@@ -8,7 +8,7 @@ awardRouter.use(login_required);
 
 //추가
 awardRouter
-	.route("/award/:user_id")
+	.route("/award/:user_id/:award_id")
 	.post(async function (req, res, next) {
 		try {
 			if (is.emptyObject(req.body)) {
@@ -17,13 +17,18 @@ awardRouter
 				);
 			}
 
-			// req (request) 에서 데이터 가져오기
-			const user_id = req.params.user_id;
+			//user validation
+			const tokenUser_id = req.currentUserId;
+			const pathUser_id = req.params.user_id;
+			if (tokenUser_id !== pathUser_id) {
+				throw new Error("인증정보가 올바르지 않습니다.");
+			}
+
 			const { title, organization, description } = req.body;
 
 			// 위 데이터를 Award db에 추가하기
 			const newAward = await AwardService.addAward({
-				user_id,
+				user_id: pathUser_id,
 				title,
 				organization,
 				description,
@@ -42,8 +47,14 @@ awardRouter
 	//수상 목록 받아오기
 	.get(async function (req, res, next) {
 		try {
-			const user_id = req.params.user_id;
-			const award = await AwardService.getAwardList({ user_id });
+			//user validation
+			const tokenUser_id = req.currentUserId;
+			const pathUser_id = req.params.user_id;
+			if (tokenUser_id !== pathUser_id) {
+				throw new Error("인증정보가 올바르지 않습니다.");
+			}
+
+			const award = await AwardService.getAwardList({ user_id: pathUser_id });
 
 			if (award.errorMessage) {
 				throw new Error(award.errorMessage);
@@ -58,8 +69,15 @@ awardRouter
 	//수정
 	.patch(async function (req, res, next) {
 		try {
-			//edu_id 추출
-			const award_id = req.body.award_id;
+			//user validation
+			const tokenUser_id = req.currentUserId;
+			const pathUser_id = req.params.user_id;
+			if (tokenUser_id !== pathUser_id) {
+				throw new Error("인증정보가 올바르지 않습니다.");
+			}
+
+			const award_id = req.params.award_id;
+
 			const { title, organization, description } = req.body ?? null;
 			const toUpdate = { title, organization, description };
 
@@ -73,23 +91,29 @@ awardRouter
 		} catch (error) {
 			next(error);
 		}
-	});
+	})
 
-awardRouter.delete("/award/:award_id", async function (req, res, next) {
-	try {
-		//id가져오기
-		const award_id = req.params.award_id;
+	.delete(async function (req, res, next) {
+		try {
+			//user validation
+			const tokenUser_id = req.currentUserId;
+			const pathUser_id = req.params.user_id;
+			if (tokenUser_id !== pathUser_id) {
+				throw new Error("인증정보가 올바르지 않습니다.");
+			}
 
-		const result = await AwardService.deleteAward({ award_id });
+			const award_id = req.params.award_id;
 
-		if (!result) {
-			throw new Error("해당 학력을 삭제할 수 없습니다.");
+			const result = await AwardService.deleteAward({ award_id });
+
+			if (!result) {
+				throw new Error("해당 학력을 삭제할 수 없습니다.");
+			}
+
+			return res.status(204).send(result);
+		} catch (error) {
+			next(error);
 		}
-
-		return res.status(204).send(result);
-	} catch (error) {
-		next(error);
-	}
-});
+	});
 
 export { awardRouter };
